Fix max/min.reduction_with_dim opinfo test for bool tensors
The tests were failing because the input tensors were getting
incorrectly promoted to int64, due to `transform_args` incorrectly
considering the `dim` argument when determining the type to promote to.

It doesn't seem like type promotion is necessary in general for max and
min, so I've switched them to `type_promotion_kind=None`.

ghstack-source-id: 7ef100256b752b75ebaba47b3ba2702e86671189
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109264