[C10d] Add PG::enableCollectivesTiming to make it dynamically enabled.
Collectives timing gates the tracking when a collective starts on device.

Currently it's enabled by set the NCCL_ENABLE_TIMING env var.

The goal of this PR is to make it possible to dynamically enable that flag so users of the PG hooks don't have to set that flag in order to have their hooks work.

The design is that once set, all new collectives will have such behavior so we track it on each Work object.

We make enableTiming_ atomic in PGNCCL to avoid races on non-TSO hardware.

To ensure consistency, we copy its value during Work construction and replace all previous usage of enableTiming_ from the PG with usages from the Work, which now has an immutable value.

ghstack-source-id: 73f27ac1971ee080a371eab7e87405a95b164397
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108814