[inductor] Add a C shim layer for libtorch
Summary: This PR adds a limited C shim layer for libtorch. The ultimate goal is to ban any direct reference to aten/c10 data structures or functions, to avoid ABI breakage by providing stable C interfaces.

To make the review and landing easier, we break the changes into several steps. In this PR, we add C interfaces for certain libtorch functions and modify the wrapper codegen to generate calls to those interfaces. There are a few major items to be addressed in future PRs:

* The AOTInductor runtime interface still takes lists of aten tensors as input and output
* The interaction with ProxyExecutor (general fallback support) needs to move away from aten tensor
* Remove all references to aten/c10 headers in the AOTInductor-generated code
* Bring up the passrate for OSS models as there are more libtorch functionalities to be supported

ghstack-source-id: 1705412fa9e8831b13480b1a4a466ab95959c93d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109022